--- src/main/java/bumblebee/xchangepass/domain/refresh/dto/RefreshTokenRequest.java
@@ -1,5 +1,0 @@
-package bumblebee.xchangepass.domain.refresh.dto;
-
-public record RefreshTokenRequest(String refreshToken) {
-
-}
