+++ src/main/java/bumblebee/xchangepass/domain/user/dto/response/UserLoginResponse.java
@@ -1,10 +1,9 @@
-package bumblebee.xchangepass.domain.user.login.dto.response;
+package bumblebee.xchangepass.domain.user.dto.response;
 
 import bumblebee.xchangepass.domain.user.entity.Role;
 import bumblebee.xchangepass.domain.user.entity.User;
 import lombok.Builder;
 
-//로그인 로직 안에서 회원 정보를 반환하는 응답 객체
 public record UserLoginResponse(
         Long userId,
         String userEmail,
