+++ src/main/java/bumblebee/xchangepass/global/security/v1/login/dto/LoginRequest.java
@@ -1,0 +1,7 @@
+package bumblebee.xchangepass.global.security.v1.login.dto;
+
+public record LoginRequest(
+        String userEmail,
+        String password
+) {
+}
