+++ src/main/java/bumblebee/xchangepass/global/security/v1/login/dto/LoginResponse.java
@@ -1,0 +1,12 @@
+package bumblebee.xchangepass.global.security.v1.login.dto;
+
+import lombok.Builder;
+
+public record LoginResponse(
+        String accessToken,
+        String refreshToken
+) {
+    @Builder
+    public LoginResponse {
+    }
+}
