+++ src/main/java/bumblebee/xchangepass/global/security/v1/refresh/dto/RefreshTokenRequest.java
@@ -1,0 +1,5 @@
+package bumblebee.xchangepass.global.security.v1.refresh.dto;
+
+public record RefreshTokenRequest(String refreshToken) {
+
+}
